work around gcc not supporting asm exceptions
I should've done this much earlier.  this only worked because I was using a custom gcc patch.  it's still a hack, but at least now it works with an unpatched gcc too.
the trick is to wrap the asm in a function marked with noipa, then the exception appears to originate from the function call instead of the asm.