rewrite mpu401_streambuf
Most importantly, we now make sure to read all incoming data before writing.
I discovered that some cards (SCC-1) will not allow writes if there is any
data present in the input FIFO, which resulted in a deadlock.

Additionally, a lot of useless code is removed, and interrupts are only
disabled when strictly necessary.